Create Parameterized Stored Procedure
This is the source code to a parameterized stored procedure that I created for a relational database in SQL Server Management Studio. In this stored procedure we can filter out the data by entering the Product ID. This stored procedure allows one to see different information about each specific product. For example, we can see the name of the product, whether that product ID is on back order, the list price for that specific item and so forth (Please refer to the source code for more information). This is simply and example of a stored procedure I created.