Create Staging Table Script
This is the source code of a staging table I created for a data mart project. The staging table script sets the primary key to clustered and adds the associated data type for each column in the table. This staging table was used for the creation of a data mart; data was extracted, transformed, and loaded from a flat file source or CSV file.